+++ common/src/main/java/mca/entity/VillagerEntityMCA.java
@@ -1156,6 +1156,8 @@ public class VillagerEntityMCA extends VillagerEntity implements VillagerLike<Vi
 
         inventory.clear();
         InventoryUtils.readFromNBT(inventory, nbt);
+
+        validateClothes();
     }
 
     @Override