+++ common/src/main/java/mca/resources/ClothingList.java
@@ -1,6 +1,7 @@
 package mca.resources;
 
 import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
 import mca.MCA;
 import mca.entity.VillagerLike;
 import mca.entity.ai.relationship.Gender;
@@ -13,12 +14,14 @@ import net.minecraft.util.registry.Registry;
 import net.minecraft.village.VillagerProfession;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.*;
+import java.io.Serializable;
+import java.util.HashMap;
+import java.util.Map;
 
 public class ClothingList extends JsonDataLoader {
     protected static final Identifier ID = MCA.locate("skins/clothing");
 
-    private final WeightedPool.Mutable<Clothing> clothing = new WeightedPool.Mutable<>(new Clothing());
+    public final HashMap<String, Clothing> clothing = new HashMap<>();
 
     private static ClothingList INSTANCE;
 
@@ -33,7 +36,7 @@ public class ClothingList extends JsonDataLoader {
 
     @Override
     protected void apply(Map<Identifier, JsonElement> data, ResourceManager manager, Profiler profiler) {
-        clothing.entries.clear();
+        clothing.clear();
 
         data.forEach((id, file) -> {
             Gender gender = Gender.byName(id.getPath().split("\\.")[0]);
@@ -43,22 +46,24 @@ public class ClothingList extends JsonDataLoader {
                 return;
             }
 
-            // adds the skins to list
-            JsonHelper.asObject(file, "root").getAsJsonObject().entrySet().forEach(entry -> {
-                int count = JsonHelper.getInt(entry.getValue().getAsJsonObject(), "count");
-                float chance = JsonHelper.getFloat(entry.getValue().getAsJsonObject(), "chance", 1.0f);
-                if (count <= 0) {
-                    return;
-                }
+            for (String key : file.getAsJsonObject().keySet()) {
+                JsonObject object = file.getAsJsonObject().get(key).getAsJsonObject();
 
-                Identifier professionIdentifier = new Identifier(entry.getKey());
-                VillagerProfession profession = Registry.VILLAGER_PROFESSION.get(professionIdentifier);
+                for (int i = 0; i < JsonHelper.getInt(object, "count", 1); i++) {
+                    String identifier = String.format(key, i);
 
-                for (int i = 0; i < count; i++) {
-                    String identifier = String.format("%s:%s/%s/%d.png", id.getNamespace(), gender.getStrName(), professionIdentifier.getPath(), i);
-                    clothing.add(new Clothing(identifier, gender, profession), chance);
+                    Clothing c = new Clothing(identifier);
+                    c.gender = gender;
+                    c.profession = JsonHelper.getString(object, "profession", null);
+                    c.chance = JsonHelper.getFloat(object, "chance", 1.0f);
+                    c.exclude = JsonHelper.getBoolean(object, "exclude", false);
+                    c.temperature = JsonHelper.getInt(object, "temperature", 0);
+
+                    if (!clothing.containsKey(identifier) || !object.has("count")) {
+                        clothing.put(identifier, c);
+                    }
                 }
-            });
+            }
         });
     }
 
@@ -68,46 +73,42 @@ public class ClothingList extends JsonDataLoader {
     public WeightedPool<String> getPool(VillagerLike<?> villager) {
         Gender gender = villager.getGenetics().getGender();
         return switch (villager.getAgeState()) {
-            case BABY, TODDLER -> getPool(gender, MCA.locate("baby"));
-            case CHILD, TEEN -> getPool(gender, MCA.locate("child"));
+            case BABY, TODDLER -> getPool(gender, MCA.locate("baby").toString());
+            case CHILD, TEEN -> getPool(gender, MCA.locate("child").toString());
             default -> getPool(gender, villager.getVillagerData().getProfession());
         };
     }
 
-    public WeightedPool<String> getPool(Gender gender, Identifier profession) {
-        return getPool(gender, Registry.VILLAGER_PROFESSION.get(profession));
+    public WeightedPool<String> getPool(Gender gender, @Nullable VillagerProfession profession) {
+        return getPool(gender, profession == null ? "minecraft:none" : Registry.VILLAGER_PROFESSION.getId(profession).toString());
     }
 
-    public WeightedPool<String> getPool(Gender gender, @Nullable VillagerProfession profession) {
-        return clothing.entries.stream()
-                .filter(c -> c.getValue().gender == Gender.NEUTRAL || gender == Gender.NEUTRAL || c.getValue().gender == gender)
-                .filter(c -> profession == null || c.getValue().profession == null || c.getValue().profession == profession)
+    public WeightedPool<String> getPool(Gender gender, @Nullable String profession) {
+        return clothing.values().stream()
+                .filter(c -> c.gender == Gender.NEUTRAL || gender == Gender.NEUTRAL || c.gender == gender)
+                .filter(c -> c.profession == null || profession == null && !c.exclude || c.profession.equals(profession))
                 .collect(() -> new WeightedPool.Mutable<>("mca:missing"),
-                        (list, entry) -> list.add(entry.getValue().identifier, entry.getWeight()),
+                        (list, entry) -> list.add(entry.identifier, entry.chance),
                         (a, b) -> {
                             a.entries.addAll(b.entries);
                         });
     }
 
-    public static class Clothing {
-        String identifier;
+    public static class Clothing implements Serializable {
+        final String identifier;
         Gender gender;
 
         @Nullable
-        VillagerProfession profession;
+        String profession;
 
-        float temperatureMin;
-        float temperatureMax;
+        float temperature;
 
-        public Clothing() {
-        }
+        float chance;
 
-        public Clothing(String identifier, Gender gender, VillagerProfession profession) {
+        boolean exclude;
+
+        public Clothing(String identifier) {
             this.identifier = identifier;
-            this.gender = gender;
-            this.profession = profession;
-            this.temperatureMin = -100.0f;
-            this.temperatureMax = 100.0f;
         }
     }
 }